+++ App.js
@@ -6,12 +6,12 @@ import { store, persistor } from './store/store';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import StartScreen from './screens/startscreen';
-import LoginScreen from './screens/LoginScreen';
-import SignUpScreen from './screens/SignupScreen';
-import SignupScreenEmail from './screens/SignupScreenEmail';
-import SignupScreenPW from './screens/SignupScreenPW';
-import SignupScreenPWCheck from './screens/SignupScreenPWCheck';
-import SignupOK from './screens/SignupOK';
+import LoginScreen from './screens/login/LoginScreen';
+import SignUpScreen from './screens/signup/SignupScreen';
+import SignupScreenEmail from './screens/signup/SignupScreenEmail';
+import SignupScreenPW from './screens/signup/SignupScreenPW';
+import SignupScreenPWCheck from './screens/signup/SignupScreenPWCheck';
+import SignupOK from './screens/signup/SignupOK';
 import MainPage from './MainPage';
 import MakeDiary from './MakeDiary';
 import ChooseDrawStyle from './ChooseDrawStyle';
