+++ screens/login/LoginScreen.js
@@ -1,12 +1,12 @@
 import React, {useState} from 'react';
 import {View, Text, StyleSheet, Image, TextInput, TouchableOpacity} from 'react-native';
-import {requestApi} from "../utils/apiSetting";
+import {requestApi} from "../../utils/apiSetting";
 import {useDispatch} from "react-redux";
-import {setAccessToken, setUserName} from "../store/reducers/user-slice";
+import {setAccessToken, setUserName} from "../../store/reducers/user-slice";
 
-const logoImage = require('../assets/icon.png');
-const emailIcon = require('../assets/email.png');
-const passwordIcon = require('../assets/password.png');
+const logoImage = require('../../assets/icon.png');
+const emailIcon = require('../../assets/email.png');
+const passwordIcon = require('../../assets/password.png');
 
 const LoginScreen = ({navigation}) => {
     const [loginForm, setLoginForm] = useState({
